model_1/airflow-requirements.txt
langchain-core
google-cloud-aiplatform
vertexai
python-Levenshtein